// src/app/interfaces/ipersona.ts
export interface Ipersona {
    id: String;
    nombres: String;
    apellido: String;
    fecha_nacimiento: String;
    nacionalidad: String;
    mail: String;
    sobre_mi: String;
    ocupacion: String;
    image_background_header: String;
    image_perfil: String;
    id_domicilio: String;
}